requirements.txt
colorlog==6.9.0
Flask==3.1.1
flask_cors==6.0.0
protobuf==3.12.4
pydantic==2.11.5
google-genai==1.19.0
eventlet==0.40.0
gevent==25.5.1
Flask-SocketIO==5.3.6
python-socketio>=5.13.0
python-engineio>=4.12.2
gunicorn==23.0.0
numpy>=2.2.6
redis~=6.2.0
flask-cors~=6.0.0